initial visualizer added. still suffers from concurrent modifications exception. first game works as expected, but following games do not check action results.

+++ src/battlechallenge/Coordinate.java
@@ -118,8 +118,8 @@ public class Coordinate implements Serializable {
 	 * @param coord the coordinate to get the euclidian distance to
 	 * @return the euclidian distance between two coordinates
 	 */
-	public double distanceTo(Coordinate coord) {
-		return Math.floor(Math.sqrt(Math.pow(this.getRow() - coord.getRow(), 2) + Math.pow(this.getCol() - coord.getCol(), 2)));
+	public int distanceTo(Coordinate coord) {
+		return (int)Math.floor(Math.sqrt(Math.pow(this.getRow() - coord.getRow(), 2) + Math.pow(this.getCol() - coord.getCol(), 2)));
 	}
 	
 } // End Coordinate Class
